Rozbicie "app.py" na mniejsze skrypty
