+++ src/app/gui/app.py
@@ -1,0 +1,11 @@
+from kivy.app import App
+from kivy.core.window import Window
+
+from app.gui.menu_and_participants_list import MainScreen
+
+
+class SumioApp(App):
+
+    def build(self):
+        Window.maximize()
+        return MainScreen()
