+++ src/app/gui/bracket.py
@@ -1,0 +1,5 @@
+from kivy.uix.boxlayout import BoxLayout
+
+
+class Bracket(BoxLayout):
+    pass
