+++ src/app/gui/load_csv.py
@@ -1,0 +1,5 @@
+from kivy.uix.boxlayout import BoxLayout
+
+
+class LoadCSV(BoxLayout):
+    pass
