+++ src/app/gui/pretty_button.py
@@ -1,0 +1,5 @@
+from kivy.uix.button import Button
+
+
+class PrettyButton(Button):
+    pass
