+++ src/app/gui/quick_start.py
@@ -1,0 +1,5 @@
+from kivy.uix.boxlayout import BoxLayout
+
+
+class QuickStart(BoxLayout):
+    pass
